requirements.txt
pandas>=2.0.0
dask>=2024.5.0
matplotlib>=3.7.0
seaborn>=0.12.0
langchain>=0.2.0
sentence-transformers>=2.6.0
faiss-cpu>=1.8.0
chromadb>=0.5.0 
gradio>=4.0.0
numpy>=1.24.0
regex>=2023.0.0
transformers>=4.30.0
torch>=2.0.0
accelerate>=0.20.0